feat: split links component
