+++ src/components/auth/auth-links/auth-links.tsx
@@ -33,24 +33,4 @@ const AuthLinks: React.FC = () => {
 	);
 };
 
-const SignInLink: React.FC = () => {
-	const { t } = useTranslation();
-
-	return (
-		<Link to={AppRoute.SIGN_IN}>
-			<Typography color="secondary.main" component="span" variant="dmSansBold">
-				{t("SignUpComponent.doYouHaveAnAccount")}
-			</Typography>
-			<Typography
-				color="primary.main"
-				component="span"
-				ml="10px"
-				variant="playfairDisplay"
-			>
-				{t("SignUpComponent.signIn")}
-			</Typography>
-		</Link>
-	);
-};
-
-export { AuthLinks, SignInLink };
+export { AuthLinks };
