+++ src/components/auth/index.ts
@@ -1,5 +1,6 @@
-export { AuthLinks, SignInLink } from "./auth-links/auth-links.tsx";
+export { AuthLinks } from "./auth-links/auth-links.tsx";
 export { GoogleButton } from "./google-button/google-button.tsx";
 export { SignInForm } from "./sign-in-form/sign-in-form.tsx";
+export { SignInLink } from "./sign-in-link/sign-in-link.tsx";
 export { SignUpForm } from "./sign-up-form/sign-up-form.tsx";
 export { StyledFormContainer } from "./styled-form-container/styles-form-container.tsx";
